apps/gemma-server/_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
datasets==3.5.0
dill==0.3.8
evaluate==0.4.3
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.1
optimum==1.24.0
packaging==25.0
pandas==2.2.3
propcache==0.3.1
protobuf==6.30.2
pyarrow==19.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
pyinstaller==6.13.0
llama_cpp_python==0.3.8
flask==3.1.0
psutil==7.0.0
beautifulsoup4==4.13.4
openai==1.79.0
python-dotenv==1.1.0